+++ rando/tourism/models.py
@@ -1,5 +1,9 @@
-from rando.core.models import JSONModel
+from rando.core.models import JSONModel, GeoJSONModel
 
 
 class DataSource(JSONModel):
     filepath = 'api/datasource/datasources.json'
+
+
+class InformationDesk(GeoJSONModel):
+    filepath = 'api/informationdesk/informationdesk.geojson'
