+++ rando/tourism/sync.py
@@ -1,8 +1,46 @@
 from django.conf import settings
 
-from rando.core.sync import InputFile
-from rando.tourism.models import DataSource
+from rando.core.sync import JSONCollection, GeoJSONCollection, reroot
+from rando.tourism import models as tourism_models
 from rando import logger
+from django.template.loader import render_to_string
+
+
+class DataSourceInputFile(JSONCollection):
+
+    def handle_record(self, record):
+        geojson_url = record['geojson_url']
+        try:
+            self.download_ressource(geojson_url, language=self.language)
+        except IOError:
+            logger.error("Could not retrieve datasource %s" % geojson_url)
+            if not settings.TOURISM_DATASOURCE_FAULT_TOLERANT:
+                raise
+
+        if self.language == settings.LANGUAGE_CODE:
+            self.download_resource(record['pictogram_url'])
+
+        return record
+
+
+class InformationDeskInputFile(GeoJSONCollection):
+
+    def handle_record(self, record):
+        record = super(InformationDeskInputFile, self).handle_record(record)
+        properties = record['properties']
+        properties['photo_url'] = reroot(properties['photo_url'])
+        properties['html'] = render_to_string('trekking/_information_desk.html',
+                                              {'desk': properties})
+
+        poitype_value = properties.pop('type')
+        properties['type'] = reroot(poitype_value, attr='pictogram')
+
+        if self.language == settings.LANGUAGE_CODE:
+            self.download_resource(properties['type']['pictogram'])
+            if properties.get('photo_url'):
+                self.download_resource(properties['photo_url'])
+
+        return record
 
 
 def sync_content_tourism(sender, **kwargs):
@@ -17,15 +55,6 @@ def sync_content_tourism(sender, **kwargs):
     logger.debug("Languages: %s" % languages)
     for language in languages:
         inputkwlang = dict(language=language, **inputkw)
-        InputFile(DataSource.filepath, **inputkwlang).pull_if_modified()
 
-        for datasource in DataSource.tmp_objects.filter(language=language).all():
-            try:
-                InputFile(datasource.geojson_url, **inputkwlang).pull_if_modified()
-            except IOError:
-                logger.error("Could not retrieve datasource %s" % datasource.geojson_url)
-                if not settings.TOURISM_DATASOURCE_FAULT_TOLERANT:
-                    raise
-
-    for datasource in DataSource.tmp_objects.filter(language=language).all():
-        InputFile(datasource.pictogram_url, **inputkw).pull_if_modified()
+        DataSourceInputFile(url=tourism_models.DataSource.filepath, **inputkwlang).pull_if_modified()
+        InformationDeskInputFile(url=tourism_models.InformationDesk.filepath, **inputkwlang).pull_if_modified()
